(add/images) reverse-oriented dockerfile
This commit adds a dockerfile for a reverse/forensic/stegano oriented
image.

I have also added code analysis tools because they could be useful as
well.

Since some reverse tools are pretty heavy to use and install, I've
decided to keep it as light as possible, so no RFID or Network tools,
but they could certainely be added as well.